n3: read histogram
Each reading call usually return a bunch of data with different data
types, different memory size and different processing needed. The idea
is to have a sort of map of this data structure to be able to
automatically read the right amount of data and correctly parse it
depending on the datatype.