n2: experimental N2 support
Main difference with N2 and new generation devices (N3, R1) is that the
new ones return OPC_BUSY (0x31) when the devices is fullfilling a
command and OPC_READY (0xF3) when it's ready for further interaction. N2
always returns 0xF3 without any ready/busy logic. At a first glance it
seems we may be able to reuse the same code, N2 is just always "ready".
Timing is mostly the same so everything should work pretty much out of
the box.