common: catch usbiss exceptions
Still not sure I want this. I mostly use the device through SPI
directly and USB seems to be a bit fragile with these power hungry
devices.
Anyway, try to catch USBIss exceptions so that continous acquisition
doesn't fail. Is this something worth to have into the lib or should
be handled in user space?